backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.6.0
requests==2.31.0
gunicorn==21.2.0
